V4.5.3 clark

See merge request !671

+++ src/main/resources/sql/V4.5.3/event_forward.sql
@@ -3,7 +3,7 @@ CREATE TABLE `event_forward` (
   `third_party_id` int(11) NOT NULL,
   `interface_id` int(11) NOT NULL,
   `user_uuid` varchar(32) COLLATE utf8mb4_unicode_ci NOT NULL,
-  `status` int(11) NOT NULL DEFAULT '1',
+  `status` tinyint NOT NULL DEFAULT '1',
   `create_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `update_at` timestamp DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   PRIMARY KEY (`id`)
